+++ CHANGELOG.md
@@ -12,6 +12,8 @@
     * Fixed `PageStorage` from detecting files as pages
     * Fixed `UserIndex` not implementing `UserCollectionInterface`
     * Fixed system templates not getting scanned [#3296](https://github.com/getgrav/grav/issues/3296)
+    * Fixed missing `onAdminAfterDelete` event call in `Flex Pages`
+>>>>>>> a8e6aedd043107a3eb2dc47bc7e2509ffb6b6db5
 
 # v1.7.9
 ## 03/19/2021